# openapi/generated.openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "p4pa-send-notification",
    "description" : "Api and Models",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
  "tags" : [ {
    "name" : "Notification",
    "description" : "the Notification API"
  }, {
    "name" : "Send",
    "description" : "the Send API"
  } ],
  "paths" : {
    "/p4pasend/send/{sendNotificationId}/preload" : {
      "post" : {
        "tags" : [ "Send" ],
        "summary" : "preload file to SEND process",
        "description" : "request to authorize preload file to SEND",
        "operationId" : "preloadSendFile",
        "parameters" : [ {
          "name" : "sendNotificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Invalid request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Notification not found"
          },
          "409" : {
            "description" : "Wrong notification status"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/p4pasend/send/{sendNotificationId}/load" : {
      "post" : {
        "tags" : [ "Send" ],
        "summary" : "load file to SEND archive process",
        "description" : "request to load file to SEND archive",
        "operationId" : "uploadSendFile",
        "parameters" : [ {
          "name" : "sendNotificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Invalid request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Notification not found"
          },
          "409" : {
            "description" : "Wrong notification status"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/p4pasend/send/{sendNotificationId}/delivery" : {
      "post" : {
        "tags" : [ "Send" ],
        "summary" : "delivery notification",
        "description" : "delivery notification",
        "operationId" : "deliveryNotification",
        "parameters" : [ {
          "name" : "sendNotificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Invalid request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Notification not found"
          },
          "409" : {
            "description" : "Wrong notification status"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/p4pasend/notification" : {
      "post" : {
        "tags" : [ "Notification" ],
        "summary" : "Create notification",
        "description" : "Create initial notification",
        "operationId" : "createSendNotification",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateNotificationRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreateNotificationResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreateNotificationResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreateNotificationResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreateNotificationResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/p4pasend/notification/{sendNotificationId}/start" : {
      "post" : {
        "tags" : [ "Notification" ],
        "summary" : "start notification process",
        "description" : "Start request to authorize preload file to SEND",
        "operationId" : "startNotification",
        "parameters" : [ {
          "name" : "sendNotificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoadFileRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StartNotificationResponse"
                }
              }
            }
          },
          "202" : {
            "description" : "Accepted and waiting for all files",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StartNotificationResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StartNotificationResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StartNotificationResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Notification or File not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StartNotificationResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Wrong notification status",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StartNotificationResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StartNotificationResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/p4pasend/send/{sendNotificationId}/status" : {
      "get" : {
        "tags" : [ "Send" ],
        "summary" : "notification status",
        "description" : "notification status",
        "operationId" : "notificationStatus",
        "parameters" : [ {
          "name" : "sendNotificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Notification not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/p4pasend/send/{sendNotificationId}/retrieveNotificationDate" : {
      "get" : {
        "tags" : [ "Send" ],
        "summary" : "retrieve notification date",
        "description" : "Retrieve notification date from send if exists",
        "operationId" : "retrieveNotificationDate",
        "parameters" : [ {
          "name" : "sendNotificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Notification not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/p4pasend/notification/{sendNotificationId}" : {
      "get" : {
        "tags" : [ "Notification" ],
        "summary" : "get send notification",
        "description" : "get send notification",
        "operationId" : "getSendNotification",
        "parameters" : [ {
          "name" : "sendNotificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "204" : {
            "description" : "NotificationData not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Notification not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendNotificationDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Notification" ],
        "summary" : "delete send notification",
        "description" : "delete send notification",
        "operationId" : "deleteSendNotification",
        "parameters" : [ {
          "name" : "sendNotificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Invalid request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Notification not found"
          },
          "409" : {
            "description" : "Wrong notification status"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "CreateNotificationResponse" : {
        "type" : "object",
        "properties" : {
          "sendNotificationId" : {
            "type" : "string"
          },
          "preloadUrl" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
      "Address" : {
        "type" : "object",
        "properties" : {
          "address" : {
            "type" : "string"
          },
          "zip" : {
            "type" : "string"
          },
          "municipality" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          }
        }
      },
      "Attachment" : {
        "type" : "object",
        "properties" : {
          "digest" : {
            "type" : "string"
          },
          "contentType" : {
            "type" : "string"
          },
          "fileName" : {
            "type" : "string"
          }
        }
      },
      "CreateNotificationRequest" : {
        "type" : "object",
        "properties" : {
          "organizationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "paProtocolNumber" : {
            "type" : "string"
          },
          "recipient" : {
            "$ref" : "#/components/schemas/Recipient"
          },
          "documents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Document"
            }
          },
          "notificationFeePolicy" : {
            "$ref" : "#/components/schemas/NotificationFeePolicyEnum"
          },
          "physicalCommunicationType" : {
            "$ref" : "#/components/schemas/PhysicalCommunicationTypeEnum"
          },
          "senderDenomination" : {
            "type" : "string"
          },
          "senderTaxId" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          },
          "paymentExpirationDate" : {
            "type" : "string",
            "format" : "date"
          },
          "taxonomyCode" : {
            "type" : "string"
          },
          "paFee" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 0
          },
          "vat" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 0
          },
          "pagoPaIntMode" : {
            "$ref" : "#/components/schemas/PagoPaIntModeEnum"
          }
        },
        "required" : [ "organizationId", "paProtocolNumber" ]
      },
      "Document" : {
        "type" : "object",
        "properties" : {
          "digest" : {
            "type" : "string"
          },
          "contentType" : {
            "type" : "string"
          },
          "fileName" : {
            "type" : "string"
          }
        }
      },
      "NotificationFeePolicyEnum" : {
        "type" : "string",
        "enum" : [ "FLAT_RATE", "DELIVERY_MODE" ]
      },
      "PagoPa" : {
        "type" : "object",
        "properties" : {
          "noticeCode" : {
            "type" : "string"
          },
          "creditorTaxId" : {
            "type" : "string"
          },
          "applyCost" : {
            "type" : "boolean"
          },
          "attachment" : {
            "$ref" : "#/components/schemas/Attachment"
          }
        }
      },
      "PagoPaIntModeEnum" : {
        "type" : "string",
        "enum" : [ "SYNC", "ASYNC", "NONE" ]
      },
      "Payment" : {
        "type" : "object",
        "properties" : {
          "pagoPa" : {
            "$ref" : "#/components/schemas/PagoPa"
          }
        }
      },
      "PhysicalCommunicationTypeEnum" : {
        "type" : "string",
        "enum" : [ "AR_REGISTERED_LETTER", "REGISTERED_LETTER_890" ]
      },
      "Recipient" : {
        "type" : "object",
        "properties" : {
          "recipientType" : {
            "$ref" : "#/components/schemas/RecipientTypeEnum"
          },
          "taxId" : {
            "type" : "string"
          },
          "denomination" : {
            "type" : "string"
          },
          "physicalAddress" : {
            "$ref" : "#/components/schemas/Address"
          },
          "payments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Payment"
            }
          }
        },
        "required" : [ "denomination", "physicalAddress", "recipientType", "taxId" ]
      },
      "RecipientTypeEnum" : {
        "type" : "string",
        "enum" : [ "PF", "PG" ]
      },
      "StartNotificationResponse" : {
        "type" : "object",
        "properties" : {
          "workFlowId" : {
            "type" : "string"
          }
        }
      },
      "LoadFileRequest" : {
        "type" : "object",
        "properties" : {
          "digest" : {
            "type" : "string"
          },
          "fileName" : {
            "type" : "string"
          }
        },
        "required" : [ "digest", "fileName" ]
      },
      "SendNotificationDTO" : {
        "type" : "object",
        "properties" : {
          "sendNotificationId" : {
            "type" : "string"
          },
          "organizationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "iun" : {
            "type" : "string"
          },
          "notificationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "payments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SendNotificationPaymentsDTO"
            }
          },
          "status" : {
            "type" : "string"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "organizationId", "payments", "sendNotificationId", "status" ]
      },
      "SendNotificationPaymentsDTO" : {
        "type" : "object",
        "properties" : {
          "debtPositionId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "navList" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "debtPositionId", "navList" ]
      }
    },
    "securitySchemes" : {
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}